small improvements and refactoring
